requirements.txt
fastapi==0.95.1
uvicorn==0.22.0

#new
SQLAlchemy==2.0.13
psycopg2-binary==2.9.6

python-dotenv==1.0.0
pydantic[email]==1.10.9
passlib[bcrypt]==1.7.4


pytest==7.4.0
httpx==0.24.1

pre-commit==3.3.3
python-jose==3.3.0
python-multipart==0.0.6

Jinja2==3.1.2
aiofiles==23.1.0